Fix dependency hell for conformance tests
